simpful/gp_fuzzy_system/evolvable_fuzzy_system.py
from simpful import FuzzySystem
import numpy as np
from copy import deepcopy

class EvolvableFuzzySystem(FuzzySystem):
    def __init__(self, *args, **kwargs):
        super().__init__(*args, **kwargs)
        self.fitness_score = None
        self.mutation_rate = 0.01  # Adjustable mutation rate for evolution

    def clone(self):
        """Creates a deep copy of the system, ensuring independent instances."""
        return deepcopy(self)

    def add_rule(self, rule):
        """Adds a new fuzzy rule to the system."""
        super().add_rules([rule])

    def mutate_rule(self):
        """Applies mutation to a randomly selected rule within the system."""
        if not self._rules:
            return  # No operation if there are no rules
        rule_index = np.random.randint(len(self._rules))
        # Access the rule condition directly assuming the structure is [condition, action]
        original_condition = self._rules[rule_index][0]
        # Perform some mutation on the condition; this is a placeholder for your mutation logic
        mutated_condition = self._mutate_rule_logic(original_condition)
        # Update the rule with the mutated condition
        self._rules[rule_index] = (mutated_condition, self._rules[rule_index][1])

    def _mutate_rule_logic(self, condition):
        """A simple example mutation that toggles 'IS' to 'IS NOT' and vice versa."""
        return condition.replace("IS", "IS NOT") if "IS" in condition else condition.replace("IS NOT", "IS")

    def crossover(self, partner_system):
        """Performs crossover between this system and another, exchanging rules."""
        if not self._rules or not partner_system._rules:
            return None, None
        crossover_point = min(len(self._rules), len(partner_system._rules)) // 2

        # Use clone to ensure that we're working with independent copies
        new_self = self.clone()
        new_partner = partner_system.clone()

        new_self._rules = new_self._rules[:crossover_point] + new_partner._rules[crossover_point:]
        new_partner._rules = new_partner._rules[:crossover_point] + new_self._rules[crossover_point:]

        return new_self, new_partner

    def evaluate_fitness(self, historical_data, predictions):
        """Calculates the fitness score based on a comparison metric like RMSE."""
        rmse = np.sqrt(np.mean((np.array(predictions) - np.array(historical_data)) ** 2))
        self.fitness_score = rmse
        return self.fitness_score


"""
Refined To-Do List for Future Enhancements:

1. Advanced Selection Mechanism:
   - Implement or integrate advanced selection mechanisms to evaluate and pick individuals according to fitness metrics such as precision, RMSE, or custom evaluation functions.

2. Crossover Operator for Rule Exchange:
   - Refine the crossover operator to efficiently exchange rules between individuals, considering rule compatibility and aiming to preserve or enhance rule effectiveness.

3. Dynamic Mutation of Rule Structure:
   - Expand mutation operations to allow for structural modifications of individuals, such as swapping variables within rules or changing the logical structure of conditions.

4. Feature Selection Through Genetic Programming:
   - Integrate mechanisms that implicitly perform feature selection during the evolutionary process, identifying and prioritizing the most informative variables for prediction.

5. Scalability and Efficiency Enhancements:
   - Assess and optimize the computational efficiency and scalability of the system, ensuring it can handle large datasets and complex rule sets.

6. Experimentation and Evaluation Framework:
    - Develop a comprehensive framework for testing and evaluating the evolved fuzzy systems across various datasets, focusing on generalization ability and predictive accuracy.

Each of these enhancements contributes directly to evolving fuzzy rule sets for data-driven solutions, aligning with the dissertation's objectives to address design problems through genetic programming. Focus on incremental development, ensuring each enhancement strengthens the system's ability to evolve and evaluate fuzzy rule sets effectively.
"""
